Fix non-editor build: SelectFlag scope and POINT narrowing
- NewUIMyInventory.cpp: `extern bool SelectFlag` was nested inside
  `#ifdef _EDITOR`, but the variable is read unconditionally a few
  lines below. Move the extern outside the editor guard.
- NewUIChatLogWindow.cpp: cast the brace-initializer expression to
  LONG to silence the narrowing-conversion error from mixing LONG
  (m_WndPos/m_WndSize) with the float RESIZING_BTN_HEIGHT constant.

+++ src/source/NewUIChatLogWindow.cpp
@@ -789,7 +789,7 @@ bool SEASON3B::CNewUIChatLogWindow::UpdateMouseEvent()
             }
         }
 
-        POINT ptResizingBtn = { m_WndPos.x, m_WndPos.y - m_WndSize.cy - RESIZING_BTN_HEIGHT };
+        POINT ptResizingBtn = { m_WndPos.x, (LONG)(m_WndPos.y - m_WndSize.cy - RESIZING_BTN_HEIGHT) };
         if (m_EventState == EVENT_NONE && false == MouseLButtonPush &&
             SEASON3B::CheckMouseIn(ptResizingBtn.x, ptResizingBtn.y, RESIZING_BTN_WIDTH, RESIZING_BTN_HEIGHT))
         {
